+++ .changeset/nice-experts-hope.md
@@ -1,0 +1,7 @@
+---
+'@clerk/clerk-js': minor
+'@clerk/types': minor
+---
+
+Introducing validatePassword for SignIn and SignUp resources
+- Validate a password based on the instance's configuration set in Password Policies in Dashboard
