+++ packages/clerk-js/src/ui/hooks/index.ts
@@ -7,7 +7,6 @@ export * from './useEnabledThirdPartyProviders';
 export * from './useLoadingStatus';
 export * from './usePassword';
 export * from './usePasswordComplexity';
-export * from './usePasswordStrength';
 export * from './usePopover';
 export * from './usePrefersReducedMotion';
 export * from './useLocalStorage';
