+++ packages/clerk-js/src/ui/utils/index.ts
@@ -19,5 +19,4 @@ export * from './roleLocalizationKey';
 export * from './getRelativeToNowDateKey';
 export * from './mergeRefs';
 export * from './createSlug';
-export * from './zxcvbn';
 export * from './passwordUtils';
