+++ packages/clerk-js/src/utils/zxcvbn.ts
